# Optimise AST rewriting