--- a/go/tools/asthelpergen/integration/rewriter.go
+++ b/go/tools/asthelpergen/integration/rewriter.go
@@ -34,16 +34,20 @@ func (a *application) apply(parent, node AST, replacer replacerFunc) {
 	case InterfaceContainer:
 	case InterfaceSlice:
 		for x, el := range n {
-			a.apply(node, el, func(newNode, container AST) {
-				container.(InterfaceSlice)[x] = newNode.(AST)
-			})
+			a.apply(node, el, func(idx int) func(AST, AST) {
+				return func(newNode, container AST) {
+					container.(InterfaceSlice)[idx] = newNode.(AST)
+				}
+			}(x))
 		}
 	case *Leaf:
 	case LeafSlice:
 		for x, el := range n {
-			a.apply(node, el, func(newNode, container AST) {
-				container.(LeafSlice)[x] = newNode.(*Leaf)
-			})
+			a.apply(node, el, func(idx int) func(AST, AST) {
+				return func(newNode, container AST) {
+					container.(LeafSlice)[idx] = newNode.(*Leaf)
+				}
+			}(x))
 		}
 	case *NoCloneType:
 	case *RefContainer:
@@ -55,14 +59,18 @@ func (a *application) apply(parent, node AST, replacer replacerFunc) {
 		})
 	case *RefSliceContainer:
 		for x, el := range n.ASTElements {
-			a.apply(node, el, func(newNode, container AST) {
-				container.(*RefSliceContainer).ASTElements[x] = newNode.(AST)
-			})
+			a.apply(node, el, func(idx int) func(AST, AST) {
+				return func(newNode, container AST) {
+					container.(*RefSliceContainer).ASTElements[idx] = newNode.(AST)
+				}
+			}(x))
 		}
 		for x, el := range n.ASTImplementationElements {
-			a.apply(node, el, func(newNode, container AST) {
-				container.(*RefSliceContainer).ASTImplementationElements[x] = newNode.(*Leaf)
-			})
+			a.apply(node, el, func(idx int) func(AST, AST) {
+				return func(newNode, container AST) {
+					container.(*RefSliceContainer).ASTImplementationElements[idx] = newNode.(*Leaf)
+				}
+			}(x))
 		}
 	case *SubImpl:
 		a.apply(node, n.inner, func(newNode, parent AST) {
@@ -73,14 +81,18 @@ func (a *application) apply(parent, node AST, replacer replacerFunc) {
 		a.apply(node, n.ASTImplementationType, replacePanic("ValueContainer ASTImplementationType"))
 	case ValueSliceContainer:
 		for x, el := range n.ASTElements {
-			a.apply(node, el, func(newNode, container AST) {
-				container.(ValueSliceContainer).ASTElements[x] = newNode.(AST)
-			})
+			a.apply(node, el, func(idx int) func(AST, AST) {
+				return func(newNode, container AST) {
+					container.(ValueSliceContainer).ASTElements[idx] = newNode.(AST)
+				}
+			}(x))
 		}
 		for x, el := range n.ASTImplementationElements {
-			a.apply(node, el, func(newNode, container AST) {
-				container.(ValueSliceContainer).ASTImplementationElements[x] = newNode.(*Leaf)
-			})
+			a.apply(node, el, func(idx int) func(AST, AST) {
+				return func(newNode, container AST) {
+					container.(ValueSliceContainer).ASTImplementationElements[idx] = newNode.(*Leaf)
+				}
+			}(x))
 		}
 	}
 	if a.post != nil && !a.post(&a.cursor) {