--- a/go/tools/asthelpergen/rewriter_gen.go
+++ b/go/tools/asthelpergen/rewriter_gen.go
@@ -133,13 +133,14 @@ func (r *rewriterGen) createReplacementMethod(container, elem types.Type, x jen.
 			return func(newnode, container AST) {
 				container.(InterfaceSlice)[idx] = newnode.(AST)
 			}
-		}
-
+		}(x)
 	*/
-	return jen.Func().Params(jen.List(jen.Id("newNode"), jen.Id("container")).Id(r.ifaceName)).Block(
-		jen.Id("container").Assert(jen.Id(types.TypeString(container, noQualifier))).Add(x).Index(jen.Id("x")).Op("=").
-			Id("newNode").Assert(jen.Id(types.TypeString(elem, noQualifier))),
-	)
+	return jen.Func().Params(jen.Id("idx").Int()).Func().Params(jen.List(jen.Id(r.ifaceName), jen.Id(r.ifaceName))).Block(
+		jen.Return(jen.Func().Params(jen.List(jen.Id("newNode"), jen.Id("container")).Id(r.ifaceName))).Block(
+			jen.Id("container").Assert(jen.Id(types.TypeString(container, noQualifier))).Add(x).Index(jen.Id("idx")).Op("=").
+				Id("newNode").Assert(jen.Id(types.TypeString(elem, noQualifier))),
+		),
+	).Call(jen.Id("x"))
 }
 
 func (r *rewriterGen) createFile(pkgName string) (string, *jen.File) {