When visiting slice elements, we were creating a closure over the index variable, like so:

```go
case *IndexHints:
	for x, el := range n.Indexes {
		a.apply(node, el, func(newNode, container SQLNode) {
			container.(*IndexHints).Indexes[x] = newNode.(ColIdent) // x is closed over
		})
	}
```

this PR changes these code snippets to be explicit about the parameter, like this:

```go
case *IndexHints:
	for x, el := range n.Indexes {
		a.apply(node, el, func(idx int) func(SQLNode, SQLNode) {
			return func(newNode, container SQLNode) {
				container.(*IndexHints).Indexes[idx] = newNode.(ColIdent)
			}
		}(x))
```

Even though the latter looks a little worse, what with the creation of a function for each element, the benchmarks show a small but significant improvement:

```
name                     old time/op  new time/op  delta
Normalize-16             4.64µs ± 1%  4.52µs ± 1%  -2.59%  (p=0.000 n=9+10)
VisitLargeExpression-16  15.7µs ± 1%  14.7µs ± 3%  -6.14%  (p=0.000 n=10+10)
```